Revert "Refactor C tests to work with Zephyr"
This reverts commit 6ce5ba26486e93d5b7696a3e23f0585932c14b16.
