+++ test/ms_vc_fff_test/ms_vc_fff_test_c/ms_vc_fff_test_c.vcxproj
@@ -20,11 +20,10 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\fff_test_c.c" />
-    <ClCompile Include="..\..\main_c.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\fff.h" />
-    <ClInclude Include="..\..\framework\c_test_framework.h" />
+    <ClInclude Include="..\..\c_test_framework.h" />
   </ItemGroup>
   <PropertyGroup>
     <CustomBuildBeforeTargets>ClCompile</CustomBuildBeforeTargets>
@@ -82,16 +81,16 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
-    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(SolutionDir)..\..\test\framework;$(IncludePath)</IncludePath>
+    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(IncludePath)</IncludePath>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
-    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(SolutionDir)..\..\test\framework;$(IncludePath)</IncludePath>
+    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(IncludePath)</IncludePath>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(SolutionDir)..\..\test\framework;$(IncludePath)</IncludePath>
+    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(IncludePath)</IncludePath>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(SolutionDir)..\..\test\framework;$(IncludePath)</IncludePath>
+    <IncludePath>$(SolutionDir)..\..\gtest;$(SolutionDir)..\..\;$(IncludePath)</IncludePath>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
